backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.2.0
pymongo==4.5.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.7
email-validator==2.1.0.post1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
